Remove types from poll.rs
